requirements.txt
fab>=1.1
redis
requests
dogpile.cache
irc
